add support for integer-typed attributes

+++ veo-rest/src/test/groovy/org/veo/core/CustomAspectPersistenceSpec.groovy
@@ -103,7 +103,7 @@ class CustomAspectPersistenceSpec extends VeoSpec {
 
         when: "add properties of type number"
 
-        cp.setProperty("my.key.3",(Integer) 10)
+        cp.setProperty("my.key.3", 10.0)
 
         assetRepository.save(assetData)
